# optimize the expression of `select 1+2`