--- a/src/functions/arithmetics/arithmetic_sub.rs
+++ b/src/functions/arithmetics/arithmetic_sub.rs
@@ -8,13 +8,13 @@ use crate::error::FuseQueryResult;
 use crate::functions::arithmetics::ArithmeticFunction;
 use crate::functions::IFunction;
 
-pub struct ArithmeticSubFunction;
+pub struct ArithmeticMinusFunction;
 
-impl ArithmeticSubFunction {
+impl ArithmeticMinusFunction {
     pub fn try_create_func(
         _ctx: FuseQueryContextRef,
         args: &[Box<dyn IFunction>],
     ) -> FuseQueryResult<Box<dyn IFunction>> {
-        ArithmeticFunction::try_create_func(DataValueArithmeticOperator::Sub, args)
+        ArithmeticFunction::try_create_func(DataValueArithmeticOperator::Minus, args)
     }
 }