mysql> select 1+1; 
+------------+
| plus(1, 1) |
+------------+
|          2 |
+------------+
1 row in set (0.01 sec)

mysql> select 1-1;
+-------------+
| minus(1, 1) |
+-------------+
|           0 |
+-------------+
1 row in set (0.01 sec)

mysql> select 1*1;
+----------------+
| multiply(1, 1) |
+----------------+
|              1 |
+----------------+
1 row in set (0.01 sec)

mysql> select 1/1;
+--------------+
| divide(1, 1) |
+--------------+
|            1 |
+--------------+
1 row in set (0.01 sec)