@sundy-li 2021-02-09
It's the style of ClickHouse but not MySQL.

Currently LGTM;